refactor(enrich): put summarization in separate folder
