+++ src/main/index.js
@@ -3,7 +3,7 @@ const _get = require("lodash/get");
 const _has = require("lodash/has");
 const extract = require("../extract");
 const enrich = require("../enrich");
-const summarizeModules = require("../enrich/summarize-modules");
+const summarizeModules = require("../enrich/summarize/summarize-modules");
 const cruiseResultSchema = require("../schema/cruise-result.schema.json");
 const meta = require("../extract/transpile/meta");
 const report = require("../report");
